GenerateNominations in canvas, type fixes
