+++ src/canvas/ManageNominations/types.ts
@@ -1,0 +1,6 @@
+// Copyright 2023 @paritytech/polkadot-staking-dashboard authors & contributors
+// SPDX-License-Identifier: GPL-3.0-only
+
+import type { Validator } from 'contexts/Validators/types';
+
+export type NewNominations = { nominations: Validator[] };
