+++ src/pages/Nominate/Active/Nominations/index.tsx
@@ -122,7 +122,6 @@ export const Nominations = ({
                       bondFor,
                       nominator,
                       nominated,
-                      nominations,
                     },
                     size: 'xl',
                   })
